# apps/coreApi/__snapshots__/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/im/conversations": {
      "post": {
        "operationId": "ImConversationController_createConversation",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a conversation or return an existing DM",
        "tags": [
          "im"
        ]
      },
      "get": {
        "operationId": "ImConversationController_listConversations",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List conversations for the current user",
        "tags": [
          "im"
        ]
      }
    },
    "/api/v1/im/conversations/search-users": {
      "get": {
        "operationId": "ImConversationController_searchUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search users within the current tenant for conversation creation",
        "tags": [
          "im"
        ]
      }
    },
    "/api/v1/im/conversations/{id}/members": {
      "get": {
        "operationId": "ImConversationController_listMembers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List members of a conversation for an authorized member",
        "tags": [
          "im"
        ]
      }
    },
...
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List users with pagination and search",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new user",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/roles": {
      "get": {
        "operationId": "RolesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List roles for a tenant",
        "tags": [
          "roles"
        ]
      },
      "post": {
        "operationId": "RolesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new role",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/roles/{id}": {
      "get": {
        "operationId": "RolesController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get role by ID",
        "tags": [
          "roles"
        ]
      },
      "patch": {
        "operationId": "RolesController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a role",
        "tags": [
          "roles"
        ]
      },
      "delete": {
        "operationId": "RolesController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a role",
        "tags": [
          "roles"
        ]
      }
    },
    "/api/v1/permissions": {
      "get": {
        "operationId": "PermissionsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all permissions",
        "tags": [
          "permissions"
        ]
      }
    },
    "/api/v1/permissions/{module}": {
      "get": {
        "operationId": "PermissionsController_findByModule",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List permissions by module",
        "tags": [
          "permissions"
        ]
      }
    },
    "/api/v1/menus": {
      "get": {
        "operationId": "MenusController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List all menus",
        "tags": [
          "menus"
        ]
      },
      "post": {
        "operationId": "MenusController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new menu item",
        "tags": [
          "menus"
        ]
      }
    },
    "/api/v1/menus/role/{roleId}": {
      "get": {
        "operationId": "MenusController_getRoleMenus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get menus assigned to a role",
        "tags": [
          "menus"
        ]
      },
      "put": {
        "operationId": "MenusController_setRoleMenus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set menus for a role",
        "tags": [
          "menus"
        ]
      }
    },
    "/api/v1/menus/user/{userId}": {
      "get": {
        "operationId": "MenusController_getUserMenus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get menus visible to a user",
        "tags": [
          "menus"
        ]
      }
    },
    "/api/v1/menus/{id}": {
      "get": {
        "operationId": "MenusController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get menu by ID",
        "tags": [
          "menus"
        ]
      },
      "patch": {
        "operationId": "MenusController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a menu item",
        "tags": [
          "menus"
        ]
      },
      "delete": {
        "operationId": "MenusController_remove",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a menu item",
        "tags": [
          "menus"
        ]
      }
    },
...
  },
  "info": {
    "title": "nodeAdmin API",
    "description": "Enterprise multi-tenant SaaS middleware platform API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}